Move move types to Move module
