+++ tests/test_castling.py
@@ -1,9 +1,9 @@
 import pytest
 
 from utahchess.board import Board
-from utahchess.castling import LONG_CASTLING, SHORT_CASTLING, get_castling_moves
+from utahchess.castling import get_castling_moves
 from utahchess.legal_moves import make_move
-from utahchess.move import Move
+from utahchess.move import LONG_CASTLING, SHORT_CASTLING, Move
 
 
 @pytest.mark.parametrize(
