+++ tests/test_en_passant.py
@@ -1,10 +1,9 @@
 import pytest
 
 from utahchess.board import Board
-from utahchess.en_passant import EN_PASSANT_MOVE, get_en_passant_moves
+from utahchess.en_passant import get_en_passant_moves
 from utahchess.legal_moves import make_move
-from utahchess.move import Move
-from utahchess.regular_move import REGULAR_MOVE
+from utahchess.move import EN_PASSANT_MOVE, REGULAR_MOVE, Move
 
 
 def test_right_side_en_passant_scenario_for_black():
